+++ src/Symfony/Component/BrowserKit/CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 ---
 
  * Add `toArray` method to `Response`
- * Add protected `getRequestClass` method to `AbstractBrowser`
+ * Add protected `createRequest` method to `AbstractBrowser`
 
 5.3
 ---